fix: fix norm
Commit type: (Choose one of the following)

Scope: (A brief description of the part of the codebase affected by the commit)

Long description (wrap at 72 characters):

Motivation:

Issue or Pull Request: (e.g., "fixes: #123" or "closes: #456" or "Issue # 789", "PR #1000")
fixes: #13
