+++ Makefile
@@ -6,7 +6,7 @@
 #    By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2023/05/16 19:51:43 by yshimoda          #+#    #+#              #
-#    Updated: 2023/06/07 16:14:54 by yshimoda         ###   ########.fr        #
+#    Updated: 2023/06/08 20:58:50 by yshimoda         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -18,13 +18,20 @@ INC_DIR		=	include
 LIBFT_DIR	=	libft
 MLX_DIR		=	minilibx-linux
 
-SRCS		=	src/main.c				\
-				src/check.c				\
-				src/read_cub_file.c		\
-				src/make_map.c		\
-				src/check_cub_file.c	\
-				src/debug.c				\
-				src/error.c
+SRCS		=	src/check_color.c				\
+				src/check_cub_file.c			\
+				src/check_file_element.c		\
+				src/check_file_name.c			\
+				src/check_texture_file.c		\
+				src/count_file_element.c		\
+				src/debug.c						\
+				src/error.c						\
+				src/input_color.c				\
+				src/input_map.c					\
+				src/input_texture_file.c		\
+				src/main.c						\
+				src/make_map.c					\
+				src/read_cub_file.c
 
 OBJS		=	$(SRCS:%.c=$(OBJ_DIR)/%.o)
 DEBUG_OBJS	=	$(SRCS:%.c=$(DEBUG_DIR)/%.o)
