+++ include/cub3d.h
@@ -6,7 +6,7 @@
 /*   By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/17 16:44:35 by yshimoda          #+#    #+#             */
-/*   Updated: 2023/06/07 16:37:15 by yshimoda         ###   ########.fr       */
+/*   Updated: 2023/06/08 20:59:45 by yshimoda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,7 +18,7 @@
 
 // remove
 # include <string.h>
-# include <stdlib.h>
+# include <ctype.h>
 
 # define COLOR_RESET	"\033[0m"
 # define COLOR_RED	"\033[31m"
@@ -104,26 +104,27 @@ enum	e_input
 	INPUT_CEILING,
 };
 
-// check
-void			exit_error(const char *s, bool perror_flag);
+void			check_color(t_game_data *data);
 void			check_cub_file(t_cub_file_node *node, t_game_data *data);
-
-// error
+void			check_file_element(t_cub_file_count *count);
 void			check_file_name(int argc, const char *argv[]);
-
-// cub_file_node
+void			check_texture_file(t_game_data *data);
+void			count_file_element(t_cub_file_node *node,
+					t_cub_file_count *count);
+void			exit_error(const char *s, bool perror_flag);
+void			input_color(t_cub_file_node *node, t_game_data *data);
+void			input_map(t_cub_file_node *node, t_game_data *data);
+void			input_texture_file(t_cub_file_node *node, t_game_data *data);
 void			free_cub_file_node(t_cub_file_node *node);
-t_cub_file_node	*make_cub_file_node(char *line);
-t_cub_file_node	*read_cub_file(const char *filename);
-
-// make_map
-size_t			lst_map_size(t_map_node *map);
+void			free_data(t_game_data *data);
+size_t			map_size(t_map_node *map);
 void			map_node_clear(t_map_node **map);
 t_map_node		*map_node_last(t_map_node *map);
 void			map_node_addback(t_map_node **map, t_map_node *next);
 t_map_node		*map_node_new(char *str);
+t_cub_file_node	*read_cub_file(const char *filename);
 
-// debug
+// debug;
 void			print_cub_file_node(t_cub_file_node *node);
 void			print_texture(t_game_data *data);
 void			print_color(t_game_data *data);
