+++ src/check_file_name.c
@@ -1,16 +1,15 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   check.c                                            :+:      :+:    :+:   */
+/*   check_file_name.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/24 14:21:57 by yshimoda          #+#    #+#             */
-/*   Updated: 2023/05/24 20:24:42 by yshimoda         ###   ########.fr       */
+/*   Updated: 2023/06/08 20:43:46 by yshimoda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-//remove
 #include "cub3d.h"
 
 void	check_file_name(int argc, const char *argv[])
