+++ src/input_texture_file.c
@@ -1,0 +1,41 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   input_texture_file.c                               :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2023/06/07 17:57:26 by yshimoda          #+#    #+#             */
+/*   Updated: 2023/06/08 19:04:34 by yshimoda         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+static char	*dup_texture_name(t_cub_file_node *node, const char *s)
+{
+	char	*texture;
+	size_t	i;
+
+	while (node)
+	{
+		if (!strncmp(node->line, s, 2))
+			break ;
+		node = node->next;
+	}
+	i = 2;
+	while (isspace(node->line[i]))
+		i++;
+	texture = strdup(&node->line[i]);
+	if (!texture)
+		exit_error("malloc error\n", false);
+	return (texture);
+}
+
+void	input_texture_file(t_cub_file_node *node, t_game_data *data)
+{
+	data->n_texture = dup_texture_name(node, "NO");
+	data->s_texture = dup_texture_name(node, "SO");
+	data->w_texture = dup_texture_name(node, "WE");
+	data->e_texture = dup_texture_name(node, "EA");
+}
