+++ src/main.c
@@ -6,42 +6,14 @@
 /*   By: yshimoda <yshimoda@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/16 19:28:11 by yshimoda          #+#    #+#             */
-/*   Updated: 2023/06/06 19:53:51 by yshimoda         ###   ########.fr       */
+/*   Updated: 2023/06/08 20:48:00 by yshimoda         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "cub3d.h"
-#include "libft.h"
+#include <stdlib.h>
 
-//void	check_map(t_map *map)
-//{
-//}
-//
-//void	check_word_exists(t_file *file, const char *str)
-//{
-//}
-//
-//void	check_identifiers_exist(t_file *file)
-//{
-//	check_word_exists(file, "NO");
-//	check_word_exists(file, "SO");
-//	check_word_exists(file, "WE");
-//	check_word_exists(file, "EA");
-//	check_word_exists(file, "F");
-//	check_word_exists(file, "C");
-//}
-//
-//void	check_file(t_file *file, t_map *map)
-//{
-//	check_identifiers_exist(file);
-//}
-//
-//void	read_file(t_file *file, const char *argv[])
-//{
-//}
-
-
-void	init_game_data(t_game_data *data)
+static void	init_game_data(t_game_data *data)
 {
 	data->map_node = NULL;
 	data->player_pos = NULL;
@@ -56,6 +28,19 @@ void	init_game_data(t_game_data *data)
 	data->c_color[0] = 0;
 	data->c_color[1] = 0;
 	data->c_color[2] = 0;
+}
+
+void	free_cub_file_node(t_cub_file_node *node)
+{
+	t_cub_file_node	*next;
+
+	while (node)
+	{
+		next = node->next;
+		free(node->line);
+		free(node);
+		node = next;
+	}
 }
 
 void	free_data(t_game_data *data)
